requirements.txt
streamlit==1.34.0
python-dotenv==1.0.1
google-generativeai==0.5.2